fix: update all imports and downstream imports for organization.module move, refs AC-1202
