+++ apps/web/src/app/admin-console/organizations/organization.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from "@angular/core";
 
-import { GroupAddEditComponent } from "../admin-console/organizations/manage/group-add-edit.component";
-import { GroupsComponent } from "../admin-console/organizations/manage/groups.component";
-import { OrganizationsRoutingModule } from "../admin-console/organizations/organization-routing.module";
-import { AccessSelectorModule } from "../admin-console/organizations/shared/components/access-selector";
+import { CoreOrganizationModule } from "../../organizations/core";
+import { SharedOrganizationModule } from "../../organizations/shared";
 
-import { CoreOrganizationModule } from "./core";
-import { SharedOrganizationModule } from "./shared";
+import { GroupAddEditComponent } from "./manage/group-add-edit.component";
+import { GroupsComponent } from "./manage/groups.component";
+import { OrganizationsRoutingModule } from "./organization-routing.module";
+import { AccessSelectorModule } from "./shared/components/access-selector";
+
 
 @NgModule({
   imports: [
