+++ apps/web/src/app/oss-routing.module.ts
@@ -27,6 +27,7 @@ import { VerifyRecoverDeleteComponent } from "../auth/verify-recover-delete.comp
 import { flagEnabled, Flags } from "../utils/flags";
 
 import { TrialInitiationComponent } from "./accounts/trial-initiation/trial-initiation.component";
+import { OrganizationModule } from "./admin-console/organizations/organization.module";
 import { AcceptFamilySponsorshipComponent } from "./admin-console/organizations/sponsorships/accept-family-sponsorship.component";
 import { FamiliesForEnterpriseSetupComponent } from "./admin-console/organizations/sponsorships/families-for-enterprise-setup.component";
 import { CreateOrganizationComponent } from "./admin-console/settings/create-organization.component";
@@ -34,7 +35,6 @@ import { SponsoredFamiliesComponent } from "./admin-console/settings/sponsored-f
 import { HomeGuard } from "./guards/home.guard";
 import { FrontendLayoutComponent } from "./layouts/frontend-layout.component";
 import { UserLayoutComponent } from "./layouts/user-layout.component";
-import { OrganizationModule } from "./organizations/organization.module";
 import { ReportsModule } from "./reports";
 import { AccessComponent } from "./send/access.component";
 import { SendComponent } from "./send/send.component";
